api/requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
drf-jwt==1.19.2
drf-schema-adapter==2.1.1
Inflector==3.0.1
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
django-feed-reader==1.0.0
psycopg2==2.9.3
